Added a bunch of functions to the user class.
All the basic add/display/edit pages should be good to go.

We might have to look into some minor updates or fixes on some pages (e.g. I want to turn the institution select list into a datatable since there are so many institutions and the user probably doesn't want to look through the list for their institution on certain pages so it would be useful to include the search feature of the datatable library.)